Add ClearSpeak prefs for Trig, Log, Functions Added tests for them
In doing this, I found that the way "Variables:" works was not powerful enough to be useful in some situations.

Variables used to stored as a Yaml value and not it is an Xpath Value, so any XPath can be stored. This necessated some changes in the Rules file (e.g, "true" -> "true()").

The variables are evaluated before the match. An optimization would be to check the match and see if any of the variables are used in it, and if not, do the evaluation after the match since it is likely the match will fail (out of hundreds of rules, only a few will match). On the other hand, only a few rules (so far) define a variable.

Changed "else_if" to "else_test" as this makes more sense. A better a different "case" command which takes pairs "if:/then:" so you don't get indenting chains for "else_test".