Initial commit to geneate Nemeth code. This gets a surprising amount of Nemeth generated correctly. It is not ready for someone to use though.
Still much work to do:
  various indicators (numeric, punctuation, ...) need work
  modifiers (munder/mover) and enclosures are not implemented
  typeface changes (should add to unicode.yaml)
  tables
Getting the indicators right is probably the hardest part.

At some point, a useful speedup up for loading would be to split the Unicode tables (there are now one for speech and one for braille) into primary and secondary parts. The primary table would be the 300 - 500 chars that are most used and the secondary one the rest of the chars. Loading unicode.yaml takes about 50ms, and doing the split will take it down to under 10ms. Maybe just 5ms. With two tables to load, that could save 90ms on startup. A flag would need to be added to SpeechRules to indicate whether the secondary table was loaded and a check that if no match was found on lookup, it should load the secondary table.

Added some basic tests for tags along with SRE tests.

The SRE tests are dubious in many cases in that some are geared towards text (footnotes, etc) and also contain bugs. They need trimming.

Probably should add a test for every rule that matters for math. The green book as about 200 rules, although many are not about math. On the other, many of the rules have subparts, so there might need to be 500 or so tests.