Added new "with: {variables: xxx, replace: yyy} construct. This solved msub, msup, and msubsup problems
To get this to work and prevent lifetime issues with Rust, it was necessary to change the 'push' command to pass the VariableDefinitions as a value, not a reference. This means they are cloned everytime they are pushed on the stack. That's not very efficient, but typically the vectory length is just 1 or 2 variables, so it isn't very big.

The lifetime problem happens because variable definitions flow into the context when pushed onto the context. This forces their lifetimes to match those of the context, but other lifetimes want them to match other things. By using values instead of references, the lifetime issues go away.

Down to one Aata "error" -- it is really a bad MathML that needs to be fixed up