The main changes are to deal with compiling to WASM for the Demo. wasm doesn't support file i/o, so I added shims for the file system calls and conditionally compile the new filesystem.rs file. For the wasm case, it imports Rules.zip which is a zipped up copy of all the rules. For "read-to-string", the file is pulled from there. There is a little more work and cleanup to do in that file.
Another thing missing in WASM is "time". Since the files can't be updated, it turns out that mostly slotted in ok. Some timings were commented out -- they could be conditionally included.

One area that they are sort of important is to generate "random" id prefixs, but I just changed to the code to the random number generater rather than the time as part of the prefix. Probably unimportant, but haven't checked.

The biggest deal is that "where_am_i_start_time" really needs a time. I hacked that to be 0 for now (since it isn't implemented), but something needs to be done. I believe that it might be possible to use the JS time function.