Two things: 1. Work on navigate.yaml    ZoomIn, ZoomInAll, ZoomOut, ZoomOutAll    MoveStart, MoveLineStart, MoveEnd, MoveLineEnd
Needed to add a support function (DistanceFromLeaf)
Needed to fix up some of the superstructure

2.  Add support to indicate braille dots for navigation
Need to add a pref
Code is pretty ugly -- I realized that it should only highlight the first braille cell, not all the cells for the character. That should simplify it some when I fix it.