Add "pronounce" TTS rule for "th" sound.
Canonicalization:
Flatten nested HTML in leaves to just be the text.
This is not necessarily the best thing to do but does simplify the code.
The alternative is to modify as_text() to return a Option<String> and all the calls. The good point is that someone could write rules for various HTML elements. For example <img> might use @alt for the speech. Of course, the flattening could do that (but doesn't right now).

Fixed/Implemented w(x,y) as function call -- that case (parenthesis with commas) seemed to have been overlooked.

Interface:
Added a way to get the value of a preference. Needed for demo to keep buttons upto date (and also so they didn't override internal changes)

Navigation:
Lots of work on navigation file. Implemented move next/previous and retooled the way the navigation words ("in numerator", etc) are spoken so they are shared.
The remaining work:
   moving in tables/2D math
   WhereAmI (did WhereAmIAll)
This resulted in changes to the code.
Also added some tests.