Changes in the way errors are handled.
The focus is on bringing the errors up to the top level so that surrounding app can decide how to handle them.

This means that print_errors becomes get_errors. A string is now returned instead of them being printed. This string is stored in many cases for future reference and checks on the error status.

The biggest change comes in reading in the files and finding the Rules dir. There is now a top-level function that must be called first to set the Rules dir. It is going to be in different locations for different apps (and that includes for the tests).

There is still more work to be done to surface more errors when applying rules.