Redid the way space is handled. Previously, it nice and simple and mostly thrown out. That's not good for UEB though.
Now mspace and mtext with only whitespace become a single non-breaking char. These are merged into adjacent mi/mn/mtext if possible so that matching rules don't need to worry about intervening spaces.

Checking for function names now does trimming because of the spaces. It's possible other places might need to do this also.

This all seems hacky, but surpringly, all of the speech tests still work.

An alternative is to maintain two versions of the global MathML instance. One for UEB with all the spaces and one with the spaces mostly removed.

A few bugs in canonicalization showed up.
One fix was that if an <mn> had a decimal pt or comma, I new consider it to be correct and don't try to do a merge.
Another was to map function names to lower cases. The UEB technical guides (weirdly) often uses a capitalized "Sin" for their examples.