One more try and getting the files to update correctly when the prefs change.
I'm not completely sold on this solution, but it works. The big problem in my mind is that the "is_up_to_date" function actually does an update of the preferences. This means the results must be consumed immediately, so various tables are cleared out and then later re-read. It seems wrong that the check causes changes.