Fixes for TTS spell command motivated by spelling chemical elements
Spell now turns multi char strings into single char ones and then does the replacement via unicode.yaml

To avoid infinite recursion, a flag was to SpeechRulesWithContext -- kind of a hack, but that's the only way to get the info
passed short of a static var

A couple of functions needed to be made public for the change.

Also, swapped the order or SSML and SAPI5 in the match statements. Maybe it is a slight speed up if it is testing one-by-one each pattern.