Lots of fixes: The main fix involves making mmultiscripts and munderover do the right thing. It passes all the tests, including many, many new ones that I added.
I also added rules for a number of special cases for mover (such as primes and dots)
There are related fixes when menclose is used

Fixes for numeric subscripts

Also, improved some indentation on some rules -- really need to go back and improve all of them.

There is one rule involving dots with a leading "." that is failing because the "." is an <mo> and isn't being recognized as a <mn> dot. Still thinking of a good fix...