Added more chars (GTM section 10)
Added a test for "|" (from Nemeth) to guess whether it is a comparison symbol ("such that") that needs spacing or absolute value.

Added a space after "." and "," when not in a number. That seems to be what the examples show.

Added G1 indicator to a few symbols. I suspect that more might need them.

Added a test for chemical elements. They emit and alternative capitalization indicator that won't allows cap word mode merging because GTM 16 suggests avoiding that for chemistry.