Added "𝐶" => "⠠" as an alternative cap indicator. This one breaks up cap word merging. It is used for chemistry where the elements should each get a cap indicator (GTM 16).
Also modified BANA rule for determining when to move to word/passage mode. Now caps don't count as a separate cell when looking in the first three cells. That cuts down on adding G1 word mode in some examples where it doesn't seem that helpful. It's a close call though...