Removed old high priority op and added a new IMPLIED_CHEMICAL_BOND high priority op that is clearer and specific to chemical bonds. Added code related to that (e.g., is_implied_chemical_bond()) and made sure that an "added" attr is part of implicit mo's that are created so they can be removed later if needed.
Cleaned up replace_children() and made it public to be used in the chemistry crate also.

If an attempt to parse as chemistry fails, the traces from that parse are removed. This includes implicit ops that were added. It is then reparsed without the chemistry artifacts. There are some shortcuts to avoid reparsing if not likely needed.

As part of that reparsing, mfenced rewrite uses a specific attr to indicate that it did its thing (doesn't need reparsing if not chemistry).

Upped the threshold for 'is chemistry'. Added a table of "all" (I hope) one and two single char element compounds so that H20 still meets the threshold, but H02 won't.

In likely_chem_formula(), set the MAYBE_CHEMISTRY values only if positive. This allows likely_chem_equation() to fire since something with operators is no longer marked as possibily being a chem formula.

Some test output needed tweaking.