Changed the way capital letters are handled. Now the AT calls and sets whether pitch should be used along with "cap" (language appropriate) should be spoken. These are controlled by the settings `CapitalLetters_Pitch` and `CapitalLetters_UseWord`.
This was done because some languages (French, Spanish, Italian, ???) speak the "cap" after the letter and that is hard to generically handle if we leave whether it is spoken up to AT interface code. Also, we don't know the word to speak (although it could be queried from MathCAT). Finally, some letters like Ω turn into "Omega", which isn't a single capital letter.

Also fixed up SAPI5 generation although I didn't test it one a SAPI voice, so it probably has some issues. At least it is plausable now...