Ran profiler to find out how to speed up loading unicode-full.yaml. Fixes #86
Biggest speedup is to have a dictionary String->XPath because building xpaths was 25% of the time. Now it is under 15%

A few more percent came from finding strings were being built in an error case but never used.

A few other tweaks were found also that contributed another few percents.