Added a hack to deal with negative numbers such as `intent='-3'`. A big problem before is that they were treated the same as literals and the "-" got dropped.
The hack will build `<negative>` appropriately when there is no preceding child. However, there is nothing appropriate to be done when we really want an "minus". Typically the preceding child is an invisible times, but by the time the intent is processed, the invisible times has been processed and generated as part of the intent tree. There could be a post processing of the intent tree, but then more magic is happening. Ugh!!!

Currently, an `mrow` with '- number' children is generated. The speech comes out as "minus" in simple cases, but it sure to be wrong in some cases.