Fix #97: make sure there is a child for elements that require one or more children.
Added data-empty-in-2D attr when there is an empty child inside a 2d structure.
We can't test for 2D in the rules because they deal with intent, so we may have "power", etc.
The attr gives us something to test.

Added a test for the attr for non-breaking space in unicode.tdl so that we hear something rather than getting "fraction over x plus y, end fraction"