Temporarily commiting code that splits `mo`s with multiple characters.
TeX creates these for relational ops such as "~~" and "==" so that the chars don't have space between them. But this can cause problems for Nemeth braille.

Splitting isn't great (maybe) as there seems to be a general advice (NFB lesson 5.9.3) to add multipurpose indicators between them even if they aren't all comparison operators. I've reached out for clarification.

Commiting so the code doesn't disappear, but I will take this out in the next commit and go with a different attempt to get it right.