Add more tests that I thought might be failures based on prelimnary comparisons with AEE.
The only one that was involved numbers after functions in enclosed lists (list_num_ind_11_a_5). That's fixed with a change to braille.rs.

However, a similar case involving angles (lesson_11_35_1) was added and that is currently broken. There seems to be some weirdness with the angle cases I need to look into.