Added new interface function ``` get_navigation_node_from_braille_position(position: usize) -> Result<(String, usize)> ``` Implements #123, but still a little buggy.
Fix bug I discovered where in Nemeth, the baseline indicator isn't highlighted when the `msup` node is the nav focus. To do this, I added a hack that hopefully doesn't get out of hand: 'b' is turned into '𝑏' to indicate this is a highlighted 'b' and the substitution has dots 7 & 8 added to it. A couple of regexps needed to change a little to handled this. This only happens if "BrailleCode" is "Nemeth" because 'b' is used for other things in UEB, etc.

I added a test that sweeps across all the positions in the quadratic formula. Nemeth passes, but now I need to work on making UEB pass that same test.