Changes related to ratio
Added the Unicode `∶` char to NemethComparisonOperators -- it is listed in the Nemeth green book. Note that canonicalization tries to figure whether colon should be that character.

Adjust the Nemeth and UEB rules to deal with that change (UEB doesn't put spaces around ration).

Change the UEB unicode.yaml entries for ratio/proportion to _not_ force G1 mode. The ICEB spec doesn't include that char.

Add "::" to canonicalization so that it produces the Unicode `∷` char.

Remove "_" from the unification of characters when not in a script position. The other (macron) chars are high and this is low, so they likely have different purposes. In particular, "_" is often used for "fill in the blank". Fixes #153.