Major refactoring of preferences with a goal to fixing updating.
The main thing is that calls to `set_preference` now set the API prefs, not the user_prefs.
This means that they will override user prefs. It also means they won't go away when a user changes a preference.

The fixes for updating added calls for invalidating both preferences and speech and calls to update. Also the initialize() function was made smarter so it can be used for doing the updates.

The tests often made internal calls to set a user preference and they don't invoke the invalidate/initialize methods, so calls were added to get the tests to work.

The update to date test was accidentally "touching" the main definitions file instead the dummy one. That's fixed.