Initial version of Spanish (and others???) "CMU" braille.
Lots more Unicode chars need to be added along with lots more tests.
There are 40 tests and 3 failures at the moment -- those all involve mmultiscripts which I haven't gotten to yet.

I'm still working my way through the manual to add/refine rules.
The manual is sometimes short on specifics, especially when it comes to when a grouping
char is needed for 2D element children.