Redid the way speech during navigation happens. This fixes #174.
Instead of just speaking the nav node (which misses context variables), the entire expression is spoken. The trick used for braille navigation highlights is extended to speech. This is done by wrapping the speech in `[[...]]` after pattern matching the nav node and then pulling the `...` out afterwards.

There is a potential problem though. The match is done on the intent tree, so the nav_node might not exist. If that is the case, we redo the speech on the mathml tree. Very likely, the reason the node id is mising is because it references some delimiters not in the intent tree.

For some things, the default `id` when `intent` is built (via `-intent:`) is wrong. So I added an optional `id` key. This is currently used when there is an `mrow` that builds the intent out of some child (e.g., `matrix`). I might not have caught all the cases.