Mostly done with 2023 update to CMU braille code.
There are 5 failures out of 208 tests. They are:
    braille::CMU::once::log_10_2_10
    braille::CMU::once::number_2_3_1
    braille::CMU::once::number_2_3_3_wiris
    braille::CMU::once::roman_numerals_appendix_2_2_1
    braille::CMU::once::units_appendix_1_3_1

The log rule requires "magically" knowing that a number represents a logarithm and has a special format.

The number rules await locale implementation.

Common roman numerals are handled approprately, the this test has a bar over the top of some that means "multiply by 1,000". The roman numeral recognition code in canonicalize.rs needs to be enhanced to catch that.

Units require special handling. No dot-3 like other multichar symbols and maybe some other issues. Awaits units implementation.

Also, there is an ignored test for color. I could do an implementation that adds some braille indicators that "there is a color change", but transcriber notes are required to indicate the colors.