This is a hack fix to a problem of when something should be sent through the unicode.yaml file and when it stays uninterpreted.
Something like `- x: "*[1]"` evaluates to a node and eventually the text comes out of it which gets run through the unicode tables. However `- x: "translate(...)"` evaluates to a string. Before this fix, the string did not get run through the table. You definitely don't want it translated if someone gives the string `- t: "foo"`, but that is also likely true for `variables: [...]` and their references (`$xxx`).

We need to be very careful about extra evaluations (such as `BrailleChars(...)` which returns things like `N⠴` which shouldn't be interpreted. We also need to be careful about infinite loops.

One solution would be to introduce a variant of `x:` Maybe `qx` (quoted x) or `xt` (execute and treat as text). However, this requires modifying hundreds of lines of rules.

The solution I implemented, which is a hack, is to introduce the idea (internally) of a quoted string. This is checked and not evaluated. BrailleChars and `$xxx` turn into quoted strings. Other things get the evaluation.

I haven't profiled the code yet. Potentially some special case checks can be made to avoid some leaf level work. For example, when stepping through some iterators that collect together a string, I see a lot going on. Most of the time, there is just a single text string so that could be checked for and all the iterator/building code can be ignored.

Fixes #199 (the 2π bug)

All tests pass after this change