Fixed #213 -- both symbol mode and word mode now allow G1 indicators to happen later in the string. This means that some alternative braille output was generated. In one case (stat_9_7_2), the result is not a listed alternative, but it is reasonable and consistent with other listed results.
Added more contractions
Improved the code to make it easier to include contractions. The main downside is that some code runs on start up that was previously a `const`, but it shouldn't be much time and only runs if UEB is used.

Need to an exceptions list (#223) and then add back the `cc` contractions (commented out due to `arccos`).