Modify `ToOrdinal` so that it takes two additional optional arguments (useFractional, isPlural) so that fractional and plural ordinals can be accessed.
This allows ToCommon() go away, but I'll do that later.

Modified the English ClearSpeak and SimpleSpeak rules to use it.

I wanted to use
```
  replace:
  - x: "*[1]"
  - x: "ToOrdinal(*[2], true, *[1]!=1)"   # extra args specify fractional ordinal and whether it is plural
```
but that causes an xpath compile error ('Invalid Argument'). For some reason `*[1]!=1` and other variants such as `*[1][.!=1]` can't be inside the function. So I resorted to using a variable and that works:
```
  variables: [IsPlural: "*[1]!=1"]
  replace:
  - x: "*[1]"
  - x: "ToOrdinal(*[2], true, $IsPlural)"   # extra args specify fractional ordinal and whether it is plural
```