As part of the change to how whitespace is handled, the rules needed to be updated.
Because whitespace is now an attribute, I added a clever rule that runs first on all tags that looks the whitespace attr. I sets a flag (which it checks), and then does a rematch on the same expr. Because of the flag, the rule won't match again. The only downside is that this match is checked first for every tag and it will rarely match. If should fail quickly though.

I also added/modified (if needed) the intent tagging for "blank" and "omission"

Also spotted an early change to the Unicode arrows for Chemistry hadn't been done to all of the files.

CMU:
* spaces for units
* omissions

Also, I removed some never used code that was leftover from starting with UEB rules/unicode.

UEB:
* spaces and omissions
* omissions

Nemeth:
* spaces and omissions

Vietnam:
This needs more work -- it has not been modified for the whitespace handling gchanges.