Another look at the CSS whitespace collapse rules and adjusted code and tests appropriately.
Whitespace is:
* collapsed in the middle of text
* if there are embedded elements, the code just grabs their text, so any whitespace around them should be collapsed
* removed at the start/end of tokens
```

As an example:
<math><mn>1</mn> <mtext>a<x>
  <y>first</y><foo></foo><y> x    second</y><y>third</y>
  </x>last  </mtext> <mi>y</mi></math>
```
should display as `1 a first x secondthird last y`.
Note that the empty `foo` element doesn't add whitespace.
Also note that ` x    second` has a single space before "x" (the `y` contents merge into the `mtext`) and the spaces after the "x" are collapsed to a single space.

Fixes #242