Added a second `definitions.yaml` file to `PreferenceManager` -- this one for braille.
Currently this is supported via a hack that does all the work of reading in the *Languages* version of the def file based on what language is associated with a braille code. This system is a hack. #12 has a suggestion of a better way to do it. But that depends upon #11 and possibly some changes to the speech rule implementation.

For now, this fixes a bug where the "Vietnam" braille code doesn't work if the speech language is not Vietnamese.