Part one of major changes to preferences and reading files.
This brings include files to the definitions.yaml file
This gets rid of the the max three locations for the file and depends on using includes.

This is not working yet. I need to fix the order of calling done.
The next big step is to move the files used out of preferences (just leaving the head) and moving the timestamps into
the rules.