Finished up redoing prefs, tracking including files, and make sure all files are up to date It passes all the tests now.
The definitions files were modified to allow hashmaps, along with existing vectors and hashset. Also "include" is now allowed. I changed the format of the defs files some.

Also, there is a separate defs file for braille -- currently they just point to the corresponding defs file in the language dir, but these can be cleaned up and some definitions moved out of the common file and into a common braille file or specific file.

Fixes: #12
Fixes: #13