Change the way `set_preference` works.
The preferences previously were partioned into API and USER ones. `set_preference` would always write to the API prefs, thereby overriding any user preference.

This caused a problem for copy/paste which needs to change the braille code. I changed this so that API and USER prefs never override one another. `set_preference` now calls `set_string_pref` instead of `set_api_pref`. This function sets the API pref if the pref name is an API pref and the USER pref if the pref name is a USER pref.

Note: user prefs, at least in NVDA, are typically changed via modifying prefs.yaml and the default is to reread that if that files changes.