Add Finnish 6-dot braille characters (unicode.yaml) (#292)
* Bring over changes made to fi-braille branch

I had a previous fork of MathCAT to which I couldn't sync to anymore. I made new fork and brought over the changes I made in the previous one. Unfortunately this erases all individual commits.

* add whitespace for arrow

* fix indent

* comment out

The product and summation are the same as Greek letters. Are these handled differently and should be included?

* add character "not tilde"

This isn't in the Finnish specification, but can be extrapolated from other structures.

* fix tilde notation

Tilde has different braille when used in geometry and when in general text. Here I opted for the geometry interpretation.

* change "double tilde" to "almost equal to"

* comment

* comment about not being in spec

* fix braille notation

* Update dummy_file_for_braille.txt

* replace p345

* replace p16 and p246

* replace markings for braille characters

* notation for factorial

* add another division symbol

* fix colon whitespace

* add arc unicode

* improve comment

* add complement

* move position of negated arrows

* add arrows and comments

* change location

* add integrals and partial derivative

* add left right arrows

* add missing notation

* Delete dummy_file_for_braille.txt

* mark translated values with T:

* replace tilde mark up with braille unicode

* don't translate parens

This is so that parens with dropped numbers works.

* add cases for exclamation point

Exclamation point is in text and in math it's factorial.