Fix #303 (crash bug in chemistry code) The problem was that in trying to unrepair the repair of splitting "OH" inside of a script, the mrow structure had changed due to canonicalization. This left no preceding children and hence the crash. The code now walks up the parent tree until there is a preceding child.
Some of the code appears to have assumed there was a single preceding child and grabbed the 0th child -- now it grabs the last child.

I also did some renaming for clarity's sake.

I added a test for this. Ideally, the code would figure out that this is (bad) chemistry. Currently, splitting "Cu" up where the "u" is the base of a superscript prevents recognition. Maybe I'll fix this in a different fix. For now, the crash is gone.