Add initial support for ":unit".
This adds several definitions in the language file: `SIPrefixes`, `SIUnits`, `SIExceptions`, `OtherUnits`, `EnglishUnits` and `PluralForm`.

The `SIExceptions` list is for collisions with other units. For example "ft" could be "femtotonnes" (femto metric tons), but is much more likely to be "feet". The initial list comes from MathPlayer's list (which looked rather buggy!).

The `PluralForms` is for exceptions where adding an "s" doesn't work. An example is "foot" -> "feet". This list is currently minimal.

Also, update to new more powerful version of "intent" rule that allows for arbitrary attributes.