Translations for bits and bytes, and also fixes for typos and such (#312)
* Finnish translations for units

Added Finnish translations for units and also grouped the units for plural forms.

* "M", or molar has no plural form in Finnish

Molar has no plural form in Finnish. This would require its own unique case that no ending letter "a" is attached to it.

* translate bites and bytes

* improve comments

* translate singular

* remove, since follows "append a" rule

* add exceptions for pluralforms

* fix typo

* fix typo: "hekto" to "hehto"

Includes fix up for the definitions and tests.

* fix typo: "kolumbi" to "kulombi"

Fixes for definitions and tests

* fix typos with "arcsecond"

* fix typo for "arcsecond" in tests

* fix test's output for baudi

* translate annum and fix tests

* fix test to reflect changes in definitions.yaml

* fix plural form for curie

* fix plural form for molaarinen in tests

This unit has no plural form in Finnish. This should throw an error if it is something else than the singular form.

* fix output for "bitti" in tests

This should throw an error if it's something else than what is specified here.