Fix #288 -- more div, grad, curl fixes
In src/canonicalize.rs, I added a hack to handle the div and curl notations that are really two characters, not one. But since the nabla could be a vector, it can't be handled by combining the characters. Instead, they get parsed into an mrow and an attribute (ACT_AS_OPERATOR) is added that says "treat this as if it were this character. The parser looks for this.

Additionally, some cleanup around nabla is also done since (at least) pandoc adds an invisible times thinking that nabla is a "mi".

Tests for cases in the issue are added as are intent rules.

A terse form for divergence is added.