Added special cases for zero matrix, indentity matrix, and diagonal matrix.
Note: the speech for diagonal matrix says "column 1, xxx, column 2, yyy, ...". Probably better to say "row nnn", and in verbose mode "row 1, column 1, xxx, ...", but that requires adding a new veriable to indicate this is a diagonal matrix, setting a default global value of false, and writing rules that check that variable then speak the right thing. The alternative to setting a variable is to have an mtd rule that does the "is this a diagonal matrix check", which is gross overkill/slow.

A bigger change would be if $XXX is not set, then the string `$XXX` is returned rather than is being an error.