Fix a nemeth bug that no one has reported and surprisingly none of the tests touch. The bug is when a nested script ends, a level indicator is emitted along with a baseline indicator; neigher should be present.
There are two regex's that deal with level and baseline indicators, but their order was wrong.

I have a test case (`nested_structure_baseline_test`) that I will submit soon, but that test also is testing braille index -> MathML and is currently failing.

An example that fails is $x^{2^n + 3^m} = 1$:
```
<math id='id-0'>
    <mrow id='id-1'>
        <msup id='id-2'>
            <mi id='id-3'>x</mi>
            <mrow id='id-4'>
                <msup id='id-5'> <mn id='id-6'>2</mn> <mi id='id-7'>n</mi> </msup>
                <mo id='id-8'>+</mo>
                <msup id='id-9'> <mn id='id-10'>3</mn>  <mi id='id-11'>m</mi> </msup>
            </mrow>
        </msup>
        <mo id='id-12'>=</mo>
        <mn id='id-13'>1</mn>
    </mrow>
</math>
```
This generates the braille `⠭⠘⠆⠘⠘⠝⠘⠬⠒⠘⠘⠍⠘⠐⠀⠨⠅⠀⠼⠂`, but the `⠘⠐` before the equals shouldn't be there. It should be `⠭⠘⠆⠘⠘⠝⠘⠬⠒⠘⠘⠍⠀⠨⠅⠀⠼⠂` because the space terminates the script.