There is one horrible, can't stay in the code, hack and two changes to handle the hack but will be needed later.
I introduced a new function that does the pattern match and can be used as value in a variable (needed by the Finnish braille translation to see if the resulting braille has whitespace, and if so, needs parens). [Not yet tested]

But for this to work, the underlying xpath needs to change. I submitted an issue on this to the xpath repo: https://github.com/shepmaster/sxd-xpath/issues/151. This commit has a horrible hack that uses `transmute` to get around the private variables in xpath. This is VERY bad. Either sxd-xpath will get fixed in the next week or two, or I will fork my own copy and add the needed functionality there.

This hack takes a Evaluation and creates a Context. A lot of copying is involved, so it is both unsafe and slow.

The proper solution is mentioned in issue: add a function to the XPath structure that acts like `evaluate` but takes an `Evaluation` (currently the XPath function takes a `Context` which internally converts that to an `Evaluation`).

The two changes that are good (but temporary/not finished) allow the pattern matching code to be reentrant. Currently these take a `Context`, but when the `evaluate` fix above is done, an `&Evaluation` should be passed and the `ContextStack` should be modified to store an Evaluation and not a Context. I didn't do this here because of the copying would likely slow MathCAT considerably. The hack only slows it when`PatternMatchResult` is called.