I made a change (https://github.com/NSoiffer/MathCAT/commit/0fc2291a8c7cfec7a572b49687fbf56c3a0dea62) a few months ago so that LiteralSpeak is used when navigating by character. However, this change didn't propogate to the last call to speak the expression, just to the navigation. This adds it to the speech so that (for example) when navigating to a paren, it will get spoken.
Initialing, I was planning to temporarily switch the pref to LiteralSpeak, but that will causes rules to be re-read. Instead, I add ":literal" which accomplishes the same thing. Potentially some ClearSpeak and SimpleSpeak rules need to check for that property, but it fixes the bugs reported.

Fixes #373

Added some tests.