Fixed navigation bugs related to "log", "ln", and "lg". The rule is now on the `mi` instead of the `mrow` so that navigation sees the pattern match on `*[1]`. This meant that the invisible function apply rule in `unicode.yaml` needed to be improved.
Fixed the rule for "unit". It had excluded the terse case. In 'en', I'm experimenting with single char units just saying their letter (e.g., "m" instead of "meter") in terse mode.

Improved the "|" rule to say "divides" in SimpleSpeak when the pattern `integer | integer` is encountered. Updated the other languages for this rule -- they had fallen behind.

Adjusted the tests accordingly.

The "trig_names" test wrongly had "of" in it. I've fixed that.