Get speech to use "ett" for 1 in front of neuter units. This includes plain numbers along with common and mixed fractions.
One needs to worry about units being in fractions (m/s), but not all cases of units are dealt with. For example, kg·m/s² is not handled.

There is a lot of duplication, but I can't see how to reduce it without some macro-like facility.

Note: '1' for a mixed fraction is still not correct.