Fixed intent's attr rule. There were several things wrong with it: * it was in an `else` clause -- it should happen regardless of `xpath-name` being present. * it was using the original mathml instead of the result of the replacement of the children in a few cases. Most notably, when it was evaluating the xpath for the attr=val clauses.
Made the regex allow for `"` and `'`.

Rewrote the DEBUG code thinking nesting `"`s were not handled correctly. It turns out the problem is that xpath doesn't allow for escaped quotes. That means that the string that is saved to print out during debugging can't have quotes in it.

The solution is to split the string and use `concat`. This involves putting the double quotes inside of single quotes. It's messy and I abandoned that.

The rewrite is a bit cleaner and allows for non-ascii characters in the DEBUG. Otherwise, no real change.