Merge nb into main branch (#387)
* Bring Norwegian translation fork into main repo (#383)

* Added Norwegian translations

* Oppdatering

* Fixed "zip in cargo and translated test for units

* Added some rules in general.yaml + some smaller changes.

* Somes changes in ClearSpeak to solve continued row. Did not work.

* Changed code for label as done for the other languages + smaller changes

* Changes of units

* Some changes in Unicode and Unicode-full

* Some clean-up and small changes.

* Translating some words from Swedish in navigate.yaml

* Changed default.yaml as done for the other languages. Now all tests pass.

* Changes in navigate.yaml

* Removing some comments in unicode-full

* Div. changes

GetBracketingIntentName to SpeakBracketingIntentName in default. Then NVDA works, but some tests fail.

Smaller changes in unicode-full.

Some changes for speaking log and ln in ClearSpeak.

* Div. changes

Replaced braces with brackets.
Replaced "structure" with "literal"
ClearSpeak-times: Copied the English version
no-say-parens: Copied the English version

* Removed some of the big numbers in definitions

* Replaced match: "." with match "count(*)=1"

* Added rule for coordinate

The rule is not tested in Norwegian

* Smaller changes

Replaced point with comma as decimal point in rules for fraction.
Copied English version of no-say-parens.

* Caused by changes in no-say-parens in both ClearSpeak and SimpleSpeak.

* Smaller changes

Replaced braces with brackets.
Added the rule "dimension-by". Not tested.

* Smaller changes

Added MatchingPause: false() at the top.
Copied English version of default, mi
Replaced structure with literal
Replaced braces with brackets

* Testing

* Div. changes

ClearSpeak: Minor clean-ups
Deafault: Minor clean-ups
General: Minor clean-ups.
Linear-algebra: Minor clean-ups
SimpleSpeak: Added short pause in the default-root-rule.
Unicode-full: Splitted some units with prefix for better pronunciation.
Unicode: Added condition in the rule for 0xB7. Now it doesn't speak "tom tom" for the test continued_row in multiline.

* Added pause

Added pause in SimpleSpeak, default, root

* Comment out intent rules

Commented out some intent rules. For example we say "divergence to" and not "divergence of" in Norwegian. These rules are in other files instead.

* Added rule with name "pause" and rule for "time"

Added the rule with name "pause"
Added a test in the rule "per-fraction". If "t" is in a fraction with units, it says "per time".

* Changed from "grunntallet" til "basen"

* Changes to improve navigation

Changed to English version for default mover, munder and munderover. It sounds better when in navigation mode.

Changed from "rottegn" to "kvadatrot" in msqrt. Sounds better for navigation.

* Minor clean-ups

Definitons.yaml: Changed the intent prefix "plus" from "positive" to "pluss". We for example say "plus infinity" and not "positive infinity".

Cleaned up some comments.

* Minor clean-ups and added test for units

* Changes to ClearSpeak

- Changed the decimal point in rules to $DecimalSeparators.
- Changed rule for units in fraction, as done for the other languages. Added test for "time".
- Removed "ClearSpeak_Fractions ='ordinal' in common-fraction rules.

* Changes to SimpleSpeak

- Changed decimal point to $DecimalSeparators
- Updated rule for units in fraction, as done for the other languages. Added test for "time"

* Changes in general

- Added rule for repeating decimal.
- Updated rule for units, as done for the other languages.
- Some changes in rule for logarithm.
- Changed the singular form of the nouns in the rule multi-line

* Changes in default

- Some changes in the literal rules for roots.
- Changed the rule literal, msub to be equal to the English.
- Some clean-ups in comments

* Minor clean-ups

- Some corrections in unicode-full
- Removing some spaces in ClearSpeak

* Added list for NeuterUnits

---------

Co-authored-by: Marthe <marthe.gjelstad@nb.no>

* Update the nb rules to match what is in 'en'. Some tests needed tweaks.

Start of special case speech for common fractions in front of units.

* Use table look up for denominators before a unit

* Tests for fractions followed by units (#385)

* Test for fractions followed by units

* 0x2225 "double vertical line" literal

Added test such that 0x2225 is spoken as "double vertical line" when navigating

* reformat a rule

* Split `NeuterUnits` into `NeuterUnits` and `NeuterUnitsWithoutPrefixes`.

* add 'Speech' arg to IsInDefinition

* Get speech to use "ett" for 1 in front of neuter units. This includes plain numbers along with common and mixed fractions.

One needs to worry about units being in fractions (m/s), but not all cases of units are dealt with. For example, kg·m/s² is not handled.

There is a lot of duplication, but I can't see how to reduce it without some macro-like facility.

Note: '1' for a mixed fraction is still not correct.

* Changed "en" to "1". MathCAT assumes "en" will be used for 1.

Adding pausing tweaks used by MathCAT.

* fixed comment

* Propogated `:unit` to `mfrac` and `msup` so that it is easier to tell what is a unit in rules.

* Fixed intent's attr rule. There were several things wrong with it:
* it was in an `else` clause -- it should happen regardless of `xpath-name` being present.
* it was using the original mathml instead of the result of the replacement of the children in a few cases. Most notably, when it was evaluating the xpath for the attr=val clauses.

Made the regex allow for `"` and `'`.

Rewrote the DEBUG code thinking nesting `"`s were not handled correctly. It turns out the problem is that xpath doesn't allow for escaped quotes. That means that the string that is saved to print out during debugging can't have quotes in it.

The solution is to split the string and use `concat`. This involves putting the double quotes inside of single quotes. It's messy and I abandoned that.

The rewrite is a bit cleaner and allows for non-ascii characters in the DEBUG. Otherwise, no real change.

* fix test that was wrong

* Added mprescript to tags that should be skipped

* Made '1' in front of mixed fraction say "ett".

Removed DEBUG(...) from all the rules.

---------

Co-authored-by: Marthe <marthe.gjelstad@nb.no>